+++ HISTORY.rst
@@ -13,6 +13,7 @@ Release History
 - python 'Typing' added
 - *Breaking* Description removed from 'create_stream'
 - locale added to Navigation
+- Certificate error messages improved
 
 **Bug Fixes**